Refactor(Core) :core:common 에 있던 파일들 :core:ui 추가 후 이동
