+++ app/build.gradle.kts
@@ -28,6 +28,7 @@ dependencies {
     implementation(project(":data"))
     implementation(project(":core:designsystem"))
     implementation(project(":core:common"))
+    implementation(project(":core:ui"))
     implementation(project(":feature:home"))
     implementation(project(":feature:detail"))
     implementation(project(":feature:favorite"))
