+++ feature/detail/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
     implementation(project(":domain"))
     implementation(project(":core:common"))
     implementation(project(":core:designsystem"))
+    implementation(project(":core:ui"))
 
     implementation(libs.landscapist.glide)
     implementation(libs.landscapist.placeholder)
