+++ feature/setting/build.gradle.kts
@@ -10,6 +10,7 @@ dependencies {
     implementation(project(":domain"))
     implementation(project(":core:common"))
     implementation(project(":core:designsystem"))
+    implementation(project(":core:ui"))
 
     implementation(libs.appcompat)
     implementation(libs.androidx.material.icons)
