+++ settings.gradle.kts
@@ -28,3 +28,4 @@ include(":core:common")
 include(":core:designsystem")
 include(":data")
 include(":domain")
+include(":core:ui")
